Change this to use the high-level method
